Event affecting other npc dialogue
* Implemented a system where a key event will affect other npc dialogues

* PlayerKeyEventManager.cs
- Manages Player's current key event states.

- Manages PlayerKeyEvent types.
Will need to add new enums to this script as development continues.

ex)
- After winning battle with Rupple, enum "Win_Rupple_StartingPoint" will be "marked" to PlayerKeyEventManager.playerKeyevents (using #markplayerevent).

- NPCs that hold certain dialogue changes due to the keyevent, when interacted it will  check if player has achieved the key event.
If so, it will change the dialogue file and remove the KeyEventDialogue class afterwards (so that it does not check this event from then on).

- If it is a one-time dialogue change,
use #rollbackdialogue tag within Dialoguemanager.cs to return to the previous dialogue.

< SIDE NOTES >
Do not change [HideInInspector] to [System.Nonserialized] within NPCController.cs.
Changing this results in an error where enemy base data vanishing within playmode.